ai-agents/tools/prebuilt-tools/duckduckgo-poc/requirements.txt
langchain
langchain-community
langgraph 
duckduckgo-search
python-dotenv
langchain-openai
openai